Manual merge of code-related changes from 64bit branch. Did NOT merge in compilation flags or changes to project layout.
git-svn-id: svn://svn.code.sf.net/p/passwordsafe/code/trunk/pwsafe/pwsafe@1704 1f79f812-37fb-46fe-a122-30589dd2bf55

+++ CompareResultsDlg.h
@@ -130,7 +130,7 @@ private:
 
 	PWScore *m_pcore0, *m_pcore1;
 
-  int m_numOnlyInCurrent, m_numOnlyInComp, m_numConflicts, m_numIdentical;
+  size_t m_numOnlyInCurrent, m_numOnlyInComp, m_numConflicts, m_numIdentical;
   int m_cxBSpace, m_cyBSpace, m_cySBar;
   int m_DialogMinWidth, m_DialogMinHeight;
   int m_DialogMaxWidth, m_DialogMaxHeight;
